fix core api config
