+++ tests/algorithm/test_tool_error_handling.py
@@ -16,8 +16,8 @@ def test_memory_tool_returns_error_result_for_unexpected_exception(monkeypatch):
     def raise_unexpected(_path, _payload):
         raise ValueError('backend payload is invalid')
 
-    monkeypatch.setattr(memory_mod, '_agentic_config', lambda: {'session_id': 'sid-1'})
-    monkeypatch.setattr(memory_mod, '_post_core_api', raise_unexpected)
+    monkeypatch.setattr(memory_mod.lazyllm, 'globals', {'agentic_config': {'session_id': 'sid-1'}})
+    monkeypatch.setattr(memory_mod, 'post_core_api', raise_unexpected)
 
     result = memory_mod.memory('memory', [{'title': 'pref', 'content': 'Remember the preference.'}])
 
@@ -31,11 +31,7 @@ def test_skill_manage_returns_error_result_for_skill_index_exception(monkeypatch
     def raise_unexpected(_base_dir):
         raise RuntimeError('skill index unavailable')
 
-    monkeypatch.setattr(
-        skill_manager_mod,
-        '_agentic_config',
-        lambda: {'session_id': 'sid-1', 'skill_fs_url': 'file:///tmp/skills'},
-    )
+    monkeypatch.setattr(skill_manager_mod.lazyllm, 'globals', {'agentic_config': {'session_id': 'sid-1'}})
     monkeypatch.setattr(skill_manager_mod, 'list_all_skill_entries', raise_unexpected)
 
     result = skill_manager_mod.skill_manage(
